feat: session plugin
