+++ fresh.config.ts
@@ -2,14 +2,14 @@
 import twindPlugin from "$fresh/plugins/twindv1.ts";
 import twindConfig from "./twind.config.ts";
 import kvOAuthPlugin from "./plugins/kv_oauth.ts";
-import protectedRoutes from "./plugins/protected_routes.ts";
+import sessionPlugin from "./plugins/session.ts";
 import errorHandling from "./plugins/error_handling.ts";
 import { FreshOptions } from "$fresh/server.ts";
 
 export default {
   plugins: [
     kvOAuthPlugin,
-    protectedRoutes,
+    sessionPlugin,
     twindPlugin(twindConfig),
     errorHandling,
   ],
