+++ routes/api/me/votes.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 the Deno authors. All rights reserved. MIT license.
 import type { Handlers } from "$fresh/server.ts";
 import { collectValues, listItemsVotedByUser } from "@/utils/db.ts";
-import { SignedInState } from "@/middleware/session.ts";
+import { SignedInState } from "@/plugins/session.ts";
 
 export const handler: Handlers<undefined, SignedInState> = {
   async GET(_req, ctx) {
